moved func to particle prop
